+++ Dev/Cpp/EffekseerRendererDX11/EffekseerRenderer/EffekseerRendererDX11.ModelLoader.cpp
@@ -52,7 +52,7 @@ Effekseer::Model* ModelLoader::Load(const char16_t* path)
 		uint8_t* data_model = new uint8_t[size_model];
 		reader->Read(data_model, size_model);
 
-		auto model = Load(data_model, size_model);
+		auto model = Load(data_model, (int32_t)size_model);
 
 		delete[] data_model;
